backend/pyproject.toml
[tool.poetry]
name = "plouf-backend"
version = "0.1.0"
description = "This is the API-Driven backend of the Plouf project."
authors = ["tristanqtn <tristan.querton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
psycopg2 = "^2.9.10"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
